requirements.txt
dj-rest-auth==2.2.5
Django==4.1.5
djangorestframework==3.14.0
requests==2.28.1
virtualenv==20.17.0
twilio==7.16.1
pyotp==2.7.0
django-allauth==0.52.0
django-rest-passwordreset==1.3.0
psycopg2==2.9.5
pillow
drf-spectacular==0.25.1